matched path: rename the directive, fixes
